Add dotenv variable parsing and versioning support
- Introduced `variables.py` for parsing and resolving environment variables using POSIX syntax.
- Added `py.typed` marker file for type hinting support.
- Updated version to 1.0.1 in `version.py`.
- Created necessary metadata files including LICENSE, METADATA, and entry points for CLI.
- Implemented RECORD file for package distribution.
- Enhanced documentation in the LICENSE and METADATA files.

+++ server/controllers/restaurant_pizza_controller.py
@@ -11,7 +11,7 @@ def create_restaurant_pizza():
     data = request.get_json()
 
     try:
-        price = int(data.get("price"))
+        price = float(data.get("price"))
         pizza_id = int(data.get("pizza_id"))
         restaurant_id = int(data.get("restaurant_id"))
     except (TypeError, ValueError):
